Bugfix: For FeatureTable `edit` test, check the QAbstractItemView.State and not isPersistentEditorOpen (#8463)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8451

# Description
In Qt 6.10.1, a bug was fixed with isPersistentEditorOpen(), where it
improperly returned True:
-  https://codereview.qt-project.org/c/qt/qtbase/+/667643

In the failing test, `edit()` is used, which actually should not result
in `isPersistentEditorOpen()` being True, because that is opened by
`isPersistentEditorOpen()`. This test was passing previously due to the
bug that was fixed in Qt 6.10.1, so now it fails. Importantly, it fails
with Pyside 6.10.1 as well.

Process: Gemini clued me into the odd check of
`isPersistentEditorOpen(idx)` when using `edit()` and then looking at
the Qt6 docs
(https://doc.qt.io/qtforpython-6/PySide6/QtWidgets/QAbstractItemView.html#PySide6.QtWidgets.QAbstractItemView.openPersistentEditor)
and searching the release notes
(https://www.qt.io/blog/qt-6.10.1-released) got me to the epiphany.

I tested locally *with Pyside6 6.10.1* which is also affected, but we
don't run pre-tests with Pyside6 -- we should! I opened:
https://github.com/napari/napari/issues/8464 .
With main the test with pyside6 6.10.1 fails and with this PR passes.