bugfix: Ensure that edge_width is accounted for when using polygon lasso (#8414)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8273

# Description
It looks like the case of `if edge and face:` and increasing number of
vertexes does not account for edge_width when setting up the meshes. In
this PR i account for it in that code block. An alternative is to remove
the early return, letting the `if edge` code take care of it, but I
suspect that would be much less performant.

Shapes tests pass for me locally and the behavior is fixed.