Fix Shapes thumbnail z ordering (#8417)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/6625

# Description
This PR removes the reversal of _z_order in ShapesList.to_colors which
is used for the rasterization of the thumbnail. This ensures that the
thumbnail reflects the canvas.
<img width="717" height="334" alt="image"
src="https://github.com/user-attachments/assets/5a68e759-2f14-4b79-b174-f05dc2efabc8"
/>

I add a test that fails on main, but passes with this PR.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>