Enh: Improve zarr reading by builtins (#8355)
# References and relevant issues
Follow-up to: https://github.com/napari/napari/pull/8268
Closes: https://github.com/napari/napari/issues/4358


# Description
The goal of this PR is to improve handling of .zarr that contain groups.
Rather than checking for .zarray, .zgroup, or zarr.json I use
`zarr.open` to open the zarr store and then get information about the
structure. Beyond that the two main changes are:

1. improving handling of groups within the root store:
A .zarr is just a container for things, which can be arrays or groups
(of images or groups) -- many users are not well versed in this nuance.
Right now our builtin reader fails if the root of the store contains
only groups and the error is rather uninformative. For example a typical
ome-zarr of a single image will be:
```
root
- /0 : this is a group with the arrays (multiscale)
- /OME : this is a group with metadata
```
Other zarr can have arbitrary structures inside, but most commonly the
root will contain a collection of images. Currently, if a zarr store is
a group, then we assume it must contain arrays to load and if it just
has groups inside, not arrays, then napari will fail to load anything.
Towards this aim, in this PR if the store is a group, I check for groups
inside and if there are groups inside **the first group (from the sorted
list) will be opened**, the user will be informed of the other groups.
Note: I don't do elaborate traversal to guess all possibilities, I just
focus on image collections -- the simplest and I suspect most common
case.

Importantly, a single-image ome-zarr is now directly openable as a zarr
"out of the box". Note that this will not invoke any of the ome
metadata, you need the plugin for that; it will just open as a
multiscale layer, which for common cases suffices. Likewise, many common
image collections will give the "expected result", e.g. whole-slide
images that are typically image collections of the data, the macro
image, and the label, will open the data correctly because the images
are typically indexed in order.

If you don't have a zarr handy, you can create a test zarr3 with two
groups with arrays using:
```
import zarr
tmp_zarr = zarr.create_group('tmp.zarr')
tmp_zarr.create_group('foo')
tmp_zarr.create_group('bar')
tmp_zarr['foo'].zeros(name='0', shape=(10, 10))
tmp_zarr['bar'].zeros(name='0', shape=(5, 5))
```
then try: `napari tmp.zarr` -- on main this will fail, with this PR it
should open the `bar` group and inform you.

2. Fix remote zarr Group handling. For the case of **remote** zarr, by
default zarr uses fsspec and HTTP, so traversing groups doesn't work:
the group_keys and array_keys return None. One needs to know the
hierarchy a priori (and/or read it from a known metadata). I added a
more informative error message for this case, instructing the user to
open a remote array -- this works with no issue and can still be useful.
Try: 
```
napari https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0062A/6001240.zarr
```
this will still fail, because it's a group, but it will be more
informative in doing so. So then try
```
napari https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0062A/6001240.zarr/2
``` 
to load an array. For an example of deeper nesting, try: 
```
napari https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0062A/6001240.zarr/labels/0/0
```  
See:
https://ome.github.io/ome-ngff-validator/?source=https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0062A/6001240.zarr
for full info. Note: these are cases that are well handled by the
napari-ome-zarr plugin, but sometimes you may want to access an array
without the plugin installed.

I added some tests for some of the new conditionals but was a bit lazy
and used Gemini for most of them.

Note I tested with envs with zarr3 and zarr2 and I tested zarr v3 files
(local and remote) and zarr v2 (local and remote).
Tests pass with both zarr2 and zarr3 envs.

---------

Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>