Enh: simple speed up of Shapes `_extent_data` (#8401)
# References and relevant issues
I've been poking around napari performance with layers with 150, 250K or
more shapes, which are not uncommon for segmented whole slide images. So
i'm (re)making a lot of viewers/layers like that. This is a simple and
small optimization that seems straightforward.

# Description
Instead of two list comprehensions, do one.
For 100K-250K shape layers this is about a 20-25% speedup of *adding a
layer*. Alas, layer creation is the lions share of the time. Still this
simple change shaves close to 2 s.
For 250K shape layer (creation takes ~21 s), with main, adding the layer
takes: 7.9-8.2 s
with this PR: 6.0-6.6 s

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>