Bugfix: update magicgui layer combobox if a layer is renamed using LayerList.renamed event (#8429)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8406
Closes: https://github.com/napari/napari/issues/4331
Alternative to: #8412 

# Description
In this PR, I add a `renamed` event to the LayerList. It's connected to
the Layer `name` event. Note: I added emitting the `index` which isn't
strictly needed, but seems useful if a `renamed` event is to exist?
Anyhow, the new LL event is then connected to `widget._reset_choices`
like the other LayerList events that are used by magicgui.
I added some simple tests for the new LL event and I added a test for
the issue behavior, that fails on main, but passes with this change.

I also tested this locally with both Image and ImageData widgets (from
napari/examples).
Renaming works for existing layers, as well as if a new layer is added
and then renamed. So I think this is working correctly.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>