Finish typing napari.layers.image (#8462)
# Description

This PR adds typing to the contstructor of `Image`, and associated other
typing to make the type checker pass. Hopefully this is a nice start to
finishing typing in the public bits of the API!

I'm sure there's room for improvement - for example, at least for now I
didn't tighten the enum-like types to more than just `str`.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>