Test on macos-15-intel without numba (#8503)
# References and relevant issues

Extracted from #8456

# Description
The llvmlite dropped support for macOS with Intel processors
(https://github.com/numba/llvmlite/pull/1298) and `llvmlite==0.46.0` is
failing to build from source on macOS.

As llvmlite is for us, transition dependency of Numba and the latest
Numba version that pins llvmlite below 0.46 is 0.62.1 then we pin Numba
for Intel macOS.

Also found that `numba` is a dependency of `glasbey` that was in
`gallery` extras. But `gallery` is in `testing` extras when it is used
only for testing examples. So remove `gallery` from `texing` extras and
modify the tox file to include gallery when testing extras.

To not explode constraint files, the test on macos-intel will be
executed without Numba.