Enh: Add label value to the labels tooltip (#8495)
# References and relevant issues
Motivated by: https://github.com/napari/napari/issues/8381
#8500 does a better job of closing that one though

# Description
Originally motived by #8381 , I struggled to sort out the caching or
other mechanism responsible for the tooltip *not* updating for a new
label because the tooltip text didn't change.
Grzegorz got this working in #8500 though, so that's the real fix.

My easy solution was to add the value to the tooltip, which then ensures
the text is different and the tooltip is displayed when hovering over a
new label even if it has the same features. It's no longer needed after
#8500 from the point of view of the issue, but, I think it's still
useful, because then you get all the info in one place: the label value
and the features. It mimics the status bar. So it's sort of a UX
enhancement I think?

I wasn't sure how to test this and couldn't find any tests for the
existing tooltip behavior, so i added a simple test to check that the
construction of the strings is correct. Open to other suggestions.