Remove shim setting and warning dialog (#8448)
# References and relevant issues
Partially addresses #8447 
Depends on https://github.com/napari/napari-plugin-manager/pull/191

# Description
This PR removes the "Use npe2 adaptor" setting and the warning dialog
introduced in 0.6.0. This means the only way to load npe1 plugins is via
the npe2 shim, and plugins that can't be auto-converted won't be loaded.
This is achieved by always overwriting the discovery step of the npe1
plugin manager with a no-op.

For most plugins this PR will have no noticeable effect. For anyone
relying on import-time behaviour, this will be a breaking change. The
best path for import-time behaviour after this PR is to rely on start-up
scripts.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>