Fix camera angles‽ (#8281)
# References and relevant issues
- see original discussion here
https://github.com/napari/napari/pull/8262#issuecomment-3303639750

# Description
In the discussion linked above we realized that the way we currently
handle camrea angles is all broken, due to a wrong conversion done deep
in vispy code (and too old to be fixed at this point). This PR rectifies
that.

In terms of mouse/ui behaviour this should be unchanged. However, the
*numbers* of the angles of the camera model should now reflect our
intuition (e.g: `(10, 0, 0)` should be a rotation of 10 degrees around
the 0th dimension).

UPDATED: The flow is like this now:
- user gives "normal" angles, corresponding to dimensions (0, 1, 2) in
napari for a normal 3D view (technically, the last 3 dimensions in
`dims.order`)
- by default, vispy understands these angles as right-handed rotations
(i.e: counterclockwise when the axis points towards the viewer). When we
flip view directions, this changes. So we introduce a sign flip here for
flipped camera views so it always behaves as right-handed rotations
- these angles are treated as `xyz` extrinsic euler, which makes it look
intuitive (first angle rotates around 0th world dimension, then 1st,
then 2nd)
- this is bundled into a rotation object (which internally uses a
quaternion, but unimportant here)
- a rotation of 90 degrees is introduced to make it so the 0th dimension
points out of the screen (this is needed because we want 2d and
"home"-3d views to match, which is NOT the case in vispy defaults, where
z points up with angles=0)
- this rotation object is then translated to what vispy actually
expects, the `yzx` euler. I'm pretty sure this is a bug, so the
conversion between the yzx euler and the quaternion looks *wrong*
because *it is*, but we retain it as is because we don't wanna break
things. It was previously hidden away by `quaternion2euler_degrees`
which we no longer use

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: napari-bot <napari-bot@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>