Workaround for Zenodo outage by downloading data from google drive.  (#8508)
# References and relevant issues

# Description
For some time, Zenodo has not been stable for download using scripts. 
I worry that it might be connected to
https://blog.zenodo.org/2025/11/25/2025-11-14-search-api-updates/

Our cache code does not work as `XDG_CACHE_HOME` was not passed through
tox, so pooch cache is in a different location than pointed to
`actions/cache`. So we cannot get data from cache.

What I have done:

1) Download  data manually (download via browser works). 
2) Upload data to Google Drive that belongs to info at napari.org 
3) Make this data publicly available via link 
4) Use `gdown` to not write manual code for confirming warning
information about a file too big for scanning it:
https://drive.usercontent.google.com/download?id=1UZEoOtHMMeJMXVoWr6IZraZnvQe8ExNJ
5) For testing, inject a custom downloader to download data from GDrive
instead of Zenodo.

This PR is not fixing docs, as it will require a PR to `conf.py`.

I do not think that it is the final solution, but it is using currently
existing resources.

As using patching code still points to the original resource.