ENH: Speedup multiple shapes removal by concat'ing and np.deleting as a batch (#8375)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8264

# Description
Add a new method to `ShapeList` to remove multiple shapes at once.
The approach is to mirror the logic of the original `remove` method, but
concat the three sets of indices and delete in one go. Then, rebuild the
array, accounting for what remains.
I used Perplexity to better understand the original remove code and help
with getting the rebuilding right after the batch removal.

Running the script from https://github.com/napari/napari/pull/3867 and
then using the GUI to select all and delete all of the Shapes is instant
now. Picking a shape and deleting appears unaffected.
Running the script from the issue, I get sub 0.1s for all three cases.
Versus main:
> So I get ~0.5 s, 2.3 s, and 62.5 s for 500, 1000, and 5000 shapes.

For 50K shapes it's ~0.4s. So the speedup is ~1000x.
All tests pass locally.

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>