Move constraints calculation to script, allow upgrade subset of packages (#8505)
# References and relevant issues

# Description

In this PR I moved code for upgrading constraints from the workflow file
to `tools/compile_constraints.sh`. This simplifies local working. You
may just run the script instead of copying action content to the
terminal.

As a follow-up, I add an option (for now only locally) to upgrade
constraints for a subset of packages just by passing them as arguments
to the script.


I think to allow this from a comment, but it is harder to test, so I
prefer to do this in a follow-up PR.

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>