Bugfix: recurse through sub-fields when making connections in EventedSettings Config (#8520)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8507

# Description
This fixes the issue by recursing through subfields. For example, in the
case of Appearance, the `highlight_thickness` is in a sub-field
`HighlightSettings`. So on main, editing the top level `them`, which is
directly in AppearanceSettings works and is saved. But the nested
settings in HIghlightSettings don't autosave.

I added an explicit test for the behavior that fails on main and passes
with this and I also tested locally with various settings to ensure
behavior wasn't changed.