Bugfix: fix erratic Shape sorting (#8408)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8389

# Description
This PR fixes the erratic shape sorting by removing the double-sort of
`triangles_z_order`. They get sorted in `_update_z_order` and then
sorted again in `_update_displayed`, which is what results in the odd
behavior when sending shapes to back or bringing to front.
Additionally, I remove the double callback on the button push which I
identified while trouble shooting the issue.

All tests pass for me locally and everything works as expected when
playing with various shapes.
'm not sure how to test this behavior because the issue is purely with
display a not the actual z_order related properties on the layer.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>